refactor: Refactor konva related methods into a folder and fix the bug where tapping draws new boxes
