+++ user_interface/src/components/BoundingBox.vue
@@ -14,7 +14,7 @@
 <script>
 import SelectModes from '@/common/enumSelectModes';
 import { mapGetters, mapMutations, mapActions } from 'vuex';
-import { OPTIMIZATION_PARAMS } from '@/common/konvaMethods';
+import { OPTIMIZATION_PARAMS } from '@/common/konva';
 import {
   DefaultBoundingBox,
   ActiveBoundingBoxFootHold,
